enhance(tests): subclass Hardfork
