+++ tests/json_infra/helpers/load_blockchain_tests.py
@@ -75,7 +75,7 @@ class BlockchainTestFixture(Fixture, FixtureTestItem):
         self.fork_name = self.test_dict["network"]
         self.add_marker(pytest.mark.fork(self.fork_name))
         self.add_marker("json_blockchain_tests")
-        self.eels_fork = FORKS[self.fork_name]
+        self.eels_fork = FORKS[self.fork_name].short_name
         test_patterns = exceptional_blockchain_test_patterns(
             self.fork_name, self.eels_fork
         )
